+++ buckaroo/buckaroo_widget.py
@@ -7,7 +7,6 @@
 """
 TODO: Add module docstring
 """
-import json
 import warnings
 import traceback
 
@@ -23,35 +22,10 @@ from .down_sample import sample
 
 from .analysis import (TypingStats, DefaultSummaryStats, ColDisplayHints)
 from .analysis_management import DfStats
-from pandas.io.json import dumps as pdumps
+
+from .serialization_utils import df_to_obj, EMPTY_DF_OBJ
 
 
-EMPTY_DF_OBJ = {'schema': {'fields': [{'name': 'index', 'type': 'string'}],
-  'primaryKey': ['index'],
-  'pandas_version': '1.4.0'},
- 'data': []}
-
-
-def dumb_table_sumarize(df):
-    """used when table_hints aren't provided.  Trests every column as a string"""
-    table_hints = {col:{'is_numeric':False}  for col in df}
-    table_hints['index'] = {'is_numeric': False} 
-    return table_hints
-
-
-def df_to_obj(df, order = None, table_hints=None):
-    if order is None:
-        order = df.columns
-    obj = json.loads(df.to_json(orient='table', indent=2, default_handler=str))
-    if table_hints is None:
-        obj['table_hints'] = json.loads(pdumps(dumb_table_sumarize(df)))
-    else:
-        obj['table_hints'] = json.loads(pdumps(table_hints))
-    fields=[{'name':'index'}]
-    for c in order:
-        fields.append({'name':c})
-    obj['schema'] = dict(fields=fields)
-    return obj
 
 
 FAST_SUMMARY_WHEN_GREATER = 1_000_000
@@ -77,7 +51,7 @@ class BuckarooWidget(DOMWidget):
     summaryDf = Dict({}).tag(sync=True)
 
     operation_results = Dict(
-        {'transformed_df':EMPTY_DF_OBJ, 'generated_py_code':'# instantiation, unused'}
+        {'transformed_df': EMPTY_DF_OBJ, 'generated_py_code':'# instantiation, unused'}
     ).tag(sync=True)
 
     dfConfig = Dict(
